Reverts name unenmanglement that was objectionable in the previous patch
commit 47bfd744177121de08fed489a5b0b1b59a1ae06a.

+++ src/library/scala/math/Numeric.scala
@@ -8,6 +8,8 @@
 
 package scala.math
 
+import language.implicitConversions
+
 /**
  * @since 2.8
  */
@@ -34,8 +36,8 @@ object Numeric {
     def fromInt(x: Int): BigInt = BigInt(x)
     def toInt(x: BigInt): Int = x.intValue
     def toLong(x: BigInt): Long = x.longValue
-    def toFloat(x: BigInt): Float = x.longValue.toFloat
-    def toDouble(x: BigInt): Double = x.longValue.toDouble
+    def toFloat(x: BigInt): Float = x.floatValue
+    def toDouble(x: BigInt): Double = x.doubleValue
   }
   implicit object BigIntIsIntegral extends BigIntIsIntegral with Ordering.BigIntOrdering
